Refactor ReAct agent, temporarily disable Canvas integration
